+++ tests/apis/agency_api/v0_x/test_devices.py
@@ -477,7 +477,7 @@ def test_device_new_api_agency_invalid(client):
 
 
 @pytest.mark.django_db
-def test_validation_error_new_api(client):
+def test_validation_error_new_api_allowed(client):
     provider = factories.Provider(id=uuid.UUID("aaaa0000-61fd-4cce-8113-81af1de90942"))
     device_id = uuid.UUID("bbbb0000-61fd-4cce-8113-81af1de90942")
     device = factories.Device(id=device_id, provider=provider)
@@ -500,5 +500,35 @@ def test_validation_error_new_api(client):
         content_type="application/json",
         **auth_header(SCOPE_AGENCY_API, provider_id=provider.id),
     )
+    assert response.status_code == 201
+    assert device.event_records.all().count() == 1
+
+
+@pytest.mark.django_db
+def test_validation_error_new_api_not_in_mapping(client):
+    provider = factories.Provider(id=uuid.UUID("aaaa0000-61fd-4cce-8113-81af1de90942"))
+    device_id = uuid.UUID("bbbb0000-61fd-4cce-8113-81af1de90942")
+    device = factories.Device(id=device_id, provider=provider)
+
+    data = {
+        "event_type": "trip_leave",
+        "event_type_reason": "off_hours",
+        "telemetry": {
+            "device_id": str(device_id),
+            "timestamp": 1_325_376_000_000,
+            "gps": {"lat": 0.0, "lng": 3.0},
+        },
+        "timestamp": 1_325_376_000_000,
+        "trip_id": None,
+    }
+
+    assert device.event_records.all().count() == 0
+    response = client.post(
+        reverse("agency:device-event", args=[device_id]),
+        data=data,
+        content_type="application/json",
+        **auth_header(SCOPE_AGENCY_API, provider_id=provider.id),
+    )
     assert response.status_code == 400
+    assert "is not in the mapping" in str(response.data)
     assert device.event_records.all().count() == 0